refactor: 레거시 관련 클래스 및 패키지 제거
