+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/AnnotationHandlerAdapter.java
@@ -1,17 +1,15 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
+package com.interface21.webmvc.servlet.mvc;
 
 import com.interface21.webmvc.servlet.HandlerAdapter;
 import com.interface21.webmvc.servlet.ModelAndView;
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
-import com.interface21.webmvc.servlet.view.JspView;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
-public final class SimpleControllerHandlerAdapter implements HandlerAdapter {
+public final class AnnotationHandlerAdapter implements HandlerAdapter {
 
     @Override
     public boolean supports(final Object handler) {
-        return handler instanceof Controller;
+        return handler instanceof HandlerExecution;
     }
 
     @Override
@@ -20,7 +18,7 @@ public final class SimpleControllerHandlerAdapter implements HandlerAdapter {
             final HttpServletResponse response,
             final Object handler
     ) throws Exception {
-        final String viewName = ((Controller) handler).execute(request, response);
-        return new ModelAndView(new JspView(viewName));
+        final HandlerExecution handlerExecution = (HandlerExecution) handler;
+        return handlerExecution.handle(request, response);
     }
 }
